task3/django/src/requirements.txt
Django==2.0.2
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
psycopg2==2.7.4
pytz==2018.3
